feat: refactor ClassDetail component and update routing for class details
